+++ R/register_all_transpilers.R
@@ -5,13 +5,13 @@ register_all_transpilers <- function() {
   transpilers_for_package("futurize::add-on", package = "base",         append_transpilers_for_future.apply)
   transpilers_for_package("futurize::add-on", package = "stats",        append_transpilers_for_future.apply)
 
-  ## Map-reduce packages (future.apply)
+  ## Map-reduce packages (pbapply)
   transpilers_for_package("futurize::add-on", package = "pbapply",      append_transpilers_for_pbapply)
 
   ## Map-reduce packages (Tidyverse)
   transpilers_for_package("futurize::add-on", package = "purrr",        append_transpilers_for_furrr)
   transpilers_for_package("futurize::add-on", package = "crossmap",     append_transpilers_for_crossmap)
-  
+
   ## Map-reduce packages (foreach)
   transpilers_for_package("futurize::add-on", package = "foreach",      append_transpilers_for_doFuture)
   transpilers_for_package("futurize::add-on", package = "plyr",         append_transpilers_for_plyr)
@@ -26,4 +26,5 @@ register_all_transpilers <- function() {
   transpilers_for_package("futurize::add-on", package = "glmnet",       append_transpilers_for_glmnet)
   transpilers_for_package("futurize::add-on", package = "lme4",         append_transpilers_for_lme4)
   transpilers_for_package("futurize::add-on", package = "tm",           append_transpilers_for_tm)
+  transpilers_for_package("futurize::add-on", package = "fwb",          append_transpilers_for_fwb)
 } ## register_all_transpilers()