+++ NEWS.md
@@ -2,9 +2,12 @@
 
 ## New Features
 
- * Add support for the **pbapply** package, e.g. `y <- pblapply(...)
-   |> futurize()`.
- 
+ * Add support for map-reduce package **pbapply**, e.g. `y <-
+   pblapply(...)  |> futurize()`.
+
+ * Add support for domain-specific package **fwb**, e.g. `b <-
+   fwb(data, statistic, R = 1000) |> futurize()`.
+
 
 # Version 0.1.0 (2026-01-18)